28.05.2019 SB: refactor to adapt for google colab
